requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Werkzeug
nltk
spacy
SpeechRecognition
pytesseract
PyPDF2
python-docx
pdfminer.six
plotly
matplotlib
WeasyPrint
python-dotenv
Pillow
numpy
pandas
transformers
torch